requirements.txt
Flask==3.0.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.102.0
pandas==2.0.3
python-dotenv==1.0.0